requirement.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2022.9.24
cffi==1.16.0
chardet==5.1.0
charset-normalizer==3.0.1
cryptography==42.0.5
dbus-python==1.3.2
distro-info==1.5+deb12u1
Django==5.0.3
google-api-core==2.18.0
google-auth==2.29.0
google-cloud==0.34.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.20.4
idna==3.3
markdown-it-py==2.1.0
mdurl==0.1.2
netifaces==0.11.0
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pycurl==7.45.2
Pygments==2.14.0
PyJWT==2.8.0
pyparsing==3.0.9
PySimpleSOAP==1.16.2
python-apt==2.6.0
python-debian==0.1.49
python-debianbts==4.0.1
PyYAML==6.0
reportbug==12.0.0
requests==2.28.1
rich==13.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
unattended-upgrades==0.1
urllib3==1.26.12
